football_fantasy_player_value_updater/Player_Value_Updater/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
lxml==4.6.3
mysql-connector-python==8.0.26
protobuf==3.17.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
Unidecode==1.2.0
urllib3==1.26.6
